requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.3
ansiconv==1.0.0
ansipants==0.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argcomplete==1.10.3
argilla==1.12.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
auto-py-to-exe==2.36.0
autocomplete==0.0.104
azure-core==1.28.0
azure-identity==1.13.0
azure-storage-blob==12.16.0
backcall==0.2.0
backoff==2.2.1
bardapi==0.1.23a0
beautifulsoup4==4.8.2
bitarray==2.7.6
black==23.7.0
bleach==6.0.0
blinker==1.6.2
bokeh==2.4.3
boto3==1.26.151
botocore==1.29.151
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chromadb==0.3.23
click==8.1.3
click-help-colors==0.9.1
click-shell==2.1
clickhouse-connect==0.5.24
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
compressed-rtf==1.0.6
contourpy==1.0.7
cryptography==41.0.1
customtkinter==5.1.3
cvlib==0.2.7
cycler==0.11.0
Cython==0.29.35
darkdetect==0.8.0
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
deep-translator==1.11.0
defusedxml==0.7.1
Deprecated==1.2.14
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@47d8ed7a1bd999719825f2e8a11d5b2828bc02c4
diffusers==0.16.1
directus-python-client==1.0.5
distlib==0.3.7
dmgbuild==1.6.1
docx2txt==0.8
ds-store==1.3.1
duckdb==0.8.0
ebcdic==1.1.1
Eel==0.16.0
effdet==0.4.1
einops==0.6.1
encodec==0.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
extract-msg==0.28.7
facexlib==0.3.0
fairscale==0.4.13
faiss-cpu==1.7.4
fastapi==0.95.2
fastjsonschema==2.16.3
filelock==3.12.0
filetype==1.2.0
filterpy==1.4.5
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
funcy==2.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
genv==1.0.0
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.2.0
google-api-core==2.11.0
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-generativeai==0.1.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
GoogleBard==1.2.1
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
h2==4.1.0
h5py==3.8.0
hnswlib==0.7.0
hpack==4.0.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.14.1
hydra-core==1.3.2
hyperframe==6.0.1
icon==0.0.5
idna==2.8
imageio==2.30.0
imaginAIry==11.0.0
IMAPClient==2.1.0
importlib-metadata==6.6.0
imutils==0.5.4
iopath==0.1.9
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.11
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.12.0
kiwisolver==1.4.4
kornia==0.6.12
langchain==0.0.225
langchainplus-sdk==0.0.20
layoutparser==0.3.4
libclang==16.0.0
lightning-utilities==0.8.0
llvmlite==0.40.0
lxml==4.9.2
lz4==4.3.2
mac-alias==2.2.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.2.0
monotonic==1.6
mpmath==1.3.0
msal==1.22.0
msal-extensions==1.0.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
np==1.0.2
numba==0.57.0
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
Office365-REST-Python-Client==2.4.2
olefile==0.46
omegaconf==2.3.0
onedrivesdk==2.0.1
open-clip-torch==2.20.0
openai==0.27.6
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.0
pdfreader==0.1.12
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.7.11
platformdirs==3.5.1
poppler-utils==0.1.0
portalocker==2.7.0
portpicker==1.2.0
posthog==3.0.1
progressbar==2.5
prometheus-client==0.16.0
prompt-toolkit==1.0.18
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.18.0
pyct==0.5.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.7.0
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.12.0
PyPDF2==3.0.1
pypdfium2==4.18.0
pyperclip==1.8.2
pyre-extensions==0.0.29
pyrsistent==0.19.3
PySocks==1.7.1
pyspnego==0.9.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-onedrive==15.10.5
python-pptx==0.6.21
pytorch-lightning==2.0.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.28.1
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.0.1
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
SharePlum==0.5.1
simplegeneric==0.8.1
six==1.12.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.8.1
SQLAlchemy==2.0.13
stack-data==0.6.2
starlette==0.27.0
streamlit==1.24.1
suno-bark @ git+https://github.com/suno-ai/bark.git@f6f2db527b13c4a3e52ed6fbac587aadc3723eb6
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
textract==1.6.5
threadpoolctl==3.1.0
tiktoken==0.4.0
timm==0.9.2
tinycss2==1.2.1
tkhtmlview==0.2.0
tkinterhtml==0.7
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
trio==0.22.0
trio-websocket==0.10.3
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.7.12
uri-template==1.2.0
urllib3==1.26.16
uvicorn==0.22.0
validators==0.20.0
virtualenv==20.24.0
virtualenv-clone==0.5.7
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.4
whichcraft==0.6.1
wrapt==1.14.1
wsproto==1.2.0
xformers==0.0.20
xlrd==1.2.0
XlsxWriter==3.1.1
yacs==0.1.8
yarl==1.9.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
zstandard==0.21.0
